Removed AspectJ based metrics for ZooKeeper (#10533)
