--- distribution/server/licenses/LICENSE-AspectJ.txt
@@ -1,64 +1,0 @@
-
-AspectJTM Compiler and Core Tools License
-
-This is a binary-only release.  Source code is available from http://eclipse.org/aspectj
-
-The Eclipse Foundation makes available all content in this distribution ("Content"). Unless otherwise
-indicated below, the Content is provided to you under the terms and conditions of the Eclipse Public
-License Version 1.0 ("EPL"). A copy of the EPL is available at http://www.eclipse.org/legal/epl-v10.html.
-For purposes of the EPL, "Program" will mean the Content.
-
-If you did not receive this Content directly from the Eclipse Foundation, the Content is being
-redistributed by another party ("Redistributor") and different terms and conditions may apply to
-your use of any object code in the Content. Check the Redistributor's license that was provided with
-the Content. If no such license exists, contact the Redistributor. Unless otherwise indicated below,
-the terms and conditions of the EPL still apply to any source code in the Content and such source
-code may be obtained at http://www.eclipse.org.
-
---- Third Party Content
-
-The Content includes items that have been sourced from third parties as set out below. If you did
-not receive this Content directly from the Eclipse Foundation, the following is provided for
-informational purposes only, and you should look to the Redistributors license for terms and
-conditions of use.
-
-BCEL v5.1
-This product contains software developed by the Apache Software Foundation (http://www.apache.org).
-
-AspectJ includes a modified version of the Apache Jakarta Byte Code Engineering Library (BCEL) v5.1.
-BCEL is available at http://jakarta.apache.org/bcel/. Source code for the modified version of BCEL
-is available at Eclipse.org in the AspectJ source tree. This code is made available under the Apache
- Software License v1.1 ASM v2.2.1
-
-AspectJ includes a binary version of ASM v2.2.1 (http://asm.objectweb.org/) The source code for ASM
-is available from the ObjectWeb download site at http://asm.objectweb.org/download/.
-
-The ASM license is available at http://asm.ow2.io/license.html. The license is also reproduced here:
-
-ASM: a very small and fast Java bytecode manipulation framework
-Copyright (c) 2000-2011 INRIA, France Telecom
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions
-are met:
-1. Redistributions of source code must retain the above copyright
-  notice, this list of conditions and the following disclaimer.
-2. Redistributions in binary form must reproduce the above copyright
-  notice, this list of conditions and the following disclaimer in the
-  documentation and/or other materials provided with the distribution.
-3. Neither the name of the copyright holders nor the names of its
-  contributors may be used to endorse or promote products derived from
-  this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
-LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
-THE POSSIBILITY OF SUCH DAMAGE.
