+++ distribution/server/src/assemble/bin.xml
@@ -130,13 +130,6 @@
       <excludes>
         <exclude>org.apache.pulsar:pulsar-functions-runtime-all</exclude>
 
-        <!-- Already included in pulsar-zookeeper instrumented jar -->
-        <exclude>org.apache.zookeeper:zookeeper</exclude>
-
-        <!-- Explicitely remove JUnit which is getting pulled in even
-             though it's set to the scope 'test' -->
-        <exclude>junit:junit</exclude>
-
         <exclude>org.projectlombok:lombok</exclude>
 
         <!-- prevent adding pulsar-functions-api-examples in lib -->
